101-print_number.c

+++ pointers_arrays_strings/101-print_number.c
@@ -1,17 +1,22 @@
 #include "main.h"
 #include <stdio.h>
 
+/**
+ * print_number - Prints a integer
+ * @n: Number to print
+ * Return: recursive
+ */
 void print_number(int n)
 {
-	int num;
+	unsigned int num;
 
 	if (n < 0)
 	{
 		_putchar('-');
-		num = -n;
+		num = -((unsigned int)n);
 	}
 	else
-		num = n;
+		num = (unsigned int) n;
 
 	if (num / 10)
 		print_number(num / 10);
